Fix Degree store keys for freki
Nodes are no longer limited to 32b

+++ community/storage-engine-util/src/main/java/org/neo4j/internal/counts/CountsKey.java
@@ -94,7 +94,7 @@ public class CountsKey {
         return first;
     }
 
-    public long second() {
+    public int second() {
         return second;
     }
